+++ WeightCore/Managers/ManagerControllerHelper.cs
@@ -10,9 +10,7 @@ namespace WeightCore.Managers
     {
         #region Design pattern "Lazy Singleton"
 
-#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         private static ManagerControllerHelper _instance;
-#pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public static ManagerControllerHelper Instance => LazyInitializer.EnsureInitialized(ref _instance);
 
         #endregion