+++ BlazorDeviceControl/CHANGELOG.md
@@ -6,8 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.6.600] - 2022-05-11
+## [0.6.610] - 2022-05-12
 ### Fixed
+- Shared.SectionWeithingFactsAggregation
+- Shared.SectionLabels
+### Описание обновления:
+- Обновлён раздел агр. взвешивания
+- Обновлён раздел этикеток
+
+## [0.6.605] - 2022-05-11
+### Fixed
+- Shared.ItemLogs
 - Shared.SectionLogs
 ### Описание обновления:
 - Обновлены раздел и страница логов